codegate/ai/bright/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
clip @ git+https://github.com/openai/CLIP.git
torch
numpy
pillow
gmpy2